Enable toggling the brush
Note that, when the brush is set to paint again, the cell where the
mouse pointer lies will highlight as well. While this behavior doesn't
arise naturally from a straightforward implementation, I found it to
be an intuitive feature to include: when I click, I want things to
happen; I expect to start painting right away.