Enable swipe-action on the slider
1. Modify 'defineGrid' to accept an existing slider button, rather
than the event. This way, I could pass in the sliderButton depending
on the kind of event that tries to access it.

2. Write 'defineGridTouch'. Since handling both touchstart and
touchmove results in the most intuitive user experience, and the same
code applies to both events, refactor that code into one function that
both events can use.

3. On my own smartphone, the 20px sliderbar height turned out to be
unusable; so I made it somewhat bigger.