Change slider's class name to something more appropriate
The original version of this project had more or less directly
followed the directions given in the Odin Project's Etch-a-Sketch
walkthrough. But since I've stopped using a prompt, the class name was
no longer apt.

Since now I'm on UI refactoring, I figure it's now a good idea to
amend this minor fault in the code.