Refactor UI: implement top-level grid layout
While much is left to be done in the way of proper sizing and styling,
the basics are in place already.

Note that care must be taken when modifying the HTML and CSS, since
the JavaScript relies on the names used in those files; an extra added
layer of nesting can completely upset the application, and can then
require some reworking in 'app.js'.