Check 'black' pen color in UI upon load
1. Explicitly denote a section of the JS code as "immediate actions",
that is, functions that get called (or events that get triggered) when
the user loads the page for the first time.

This code uses a lot of event listeners; there's very little that
happens "right away" (e.g., via function invocation), as in some other
kinds of programs. So I want to therefore make that clear distinction.