Use Flexbox for top-level layout
There is still a major problem, one that arguably transcends whether
Flexbox or Grid is used.

When rotating the screen on mobile devices, since a new grid isn't
being created, the cell offsets get ruined, and I don't know yet how
to, for example, keep the Etch-a-Sketch grid size constant, relative
to the device.