Make slider respond to dragging instead of mouseover
Rename 'defineGrid' to 'doSliderButton', since dragging the
slider widget only updates the position of the widget; it doesn't
define a new grid for every touched slider position.

This solves partially the slowness problem when a grid with a lot of
squares is requested: on letting go of the widget on a high grid
setting, the user need only wait a bit for the landed-on grid setting
to take effect, rather than become distracted by various large grids
queueing to form in sequence.