Eliminate paintOn flag and forcePaint argument
I don't remember exactly why I had these in the first place, but I'm
pretty sure it had to do with the way painting was being toggled by
the mouse previously.

At this point, the application relies on the more intuitive,
mobile-device-inspired behavior of click-and-drag painting. Hence, no
more 'paintOn' toggling is needed, nor do we need to patch this leaky
abstraction with a 'forcePaint' argument to the painter function.