Avoid consecutive firing of 'constructGrid'
The event behavior we have is the ideal one, and I'd rather not change
it. Hence, I've patched this function to simply not repeat-fire if
it's being asked to reconstruct a grid with the same dimensions as the
previous one (as such a grid would be the exact same grid). This is,
at any rate, a sensible precaution (though one that normally wouldn't
be needed in the normal course of things.)