Change CSS to reflect initial background color setting
I also edited the HTML to reflect the fact that the button is no
longer a "paint white" setting, but a "paint with current background
color" setting.

The only thing remaining is to change the cells that possess the
current background color, to have the new background color when so
triggered.