Fix rgb/hex issue in equlity comparison for background change
The fix is to immediately convert the color picker's hex format to the
program's convention of using rgb format. While there are more
orthodox, "programmatic" ways of performing the conversion, it struck
me that the bug went undetected in the first place, _precisely because
the DOM converts hex to RGB when you set a DOM element's inline color
style._ However, the equality comparsion that checks a cell's
background color of course fails later, since for example #ffffff !==
rgb(255, 255, 255). So this commit remedies that, using the DOM's own
methods.