Treat custom foreground color the same way as background color
Unfortunately, this refactor is needed as a result of an annoyance
regarding the HTML5 color picker: there are no "set" or "cancel"
events. That is, I can only detect whether the color has been
literally changed.

This means I can't use it to persist a custom color value for me. I
started to rely on this feature as I was redesigning the application
logo, but soon ran into bugs. So, I have to do it myself: I've
therefore separated the color picker from the corresponding pen
setting.