WIP: introduce lightening and darkening
To avoid code duplication and still keep the event handler code
simpler (by not having to explicitly pass the 'darken' parameter), I'm
using a factory to create the lightener and darkener functions.

This commit is technically broken since switching to another button
doesn't change the pen function back to 'usePen', but I wanted to save
my progress here with this WIP commit.